Switch metric/segment `is_active` to `archived`
This column name and logic matches what is done currently for other
entities in the model. This commit also adds a `archived` column for
pulses/alerts to make it more consistent with the other entities. The
`archived` column for Pulses is just included in responses currently,
there's no plumbing to populate it, filter by it etc.